third_party/blink/web_tests/platform/generic/http/tests/devtools/console/console-viewport-indices-expected.txt

Verifies viewport's visible and active message ranges.


Running: testEmptyViewport
Expected and actual visible ranges match

Running: testFirstLastVisibleIndices
Logging 100 Normal messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match

Running: testMultilineMessages
Logging 100 Multiline messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match

Running: testSlightlyBiggerMessages
Logging 100 SlightlyBigger messages
Force item to be first: 0
Expected and actual visible ranges match
Force item to be first: 1
Expected and actual visible ranges match
Scroll a bit down: 15px
Expected and actual visible ranges match
Force item to be last: 50
Expected and actual visible ranges match
Force item to be last: 99
Expected and actual visible ranges match
